Agentic_AI/auto_audit/requirements.txt
Flask>=2.2.0
pandas>=2.1.0
google-generativeai>=0.7.0
SQLAlchemy>=2.0.0
snowflake-connector-python==3.11.0
snowflake-sqlalchemy>=1.5.0
cryptography==41.0.3